Remove fee limit related logic from ReceivableBTCBox
Extract FeeMessage to it's own file
